Added trans3d_translate to transformer.h|cpp

+++ bezier.cpp
@@ -47,7 +47,7 @@ void bezierPoint(const float u, Pt3 * pt, const int nControls, const Pt3 *contro
 	
 	// Add in influence of each control point
 	for (k = 0; k < nControls; k++){
-		blend = c[k] * powf(u, k) *powf(1 - u, n - k);
+		blend = (float)c[k] * powf(u, k) *powf(1 - u, n - k);
 		pt->x += controls[k].x * blend;
 		pt->y += controls[k].y * blend;
 		pt->z += controls[k].z * blend;
